% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_odd.R
\name{is_odd}
\alias{is_odd}
\title{Check whether a number is even or odd}
\usage{
is_odd(number)
}
\arguments{
\item{number}{A numeric vector.}
}
\value{
Returns a string, indicating whether a given number in \code{number} are even or odd.
}
\description{
This function will take a numeric vector \code{number} and performs a modulo to get the division remainder of 2, and compares it to zero. Returns \code{even} or \code{odd}.
}
\examples{
is_odd(1)
is_odd(2)

}
